fix: stabilize task ports and PR checks
