+++ src/specify_cli/agent_tasks_ports.py
@@ -1,0 +1,381 @@
+"""Injected capability ports for the ``agent tasks`` command surface (WP02).
+
+Co-designed capability boundary for the ``tasks.py`` de-god program
+(``tasks-py-degod-01KWF08S``) and reused by Wave 2. The four ports are
+**stratified** (research D2):
+
+* **Program-reference ports** (Wave 2 reuses these):
+  - :class:`FsReader` — coordination **READ** authority (C-001). Never conflates
+    with the write authority.
+  - :class:`CoordCommitRouter` — coordination **WRITE** authority, exposing
+    **two capability methods over two structurally disjoint real seams**
+    (:meth:`~CoordCommitRouter.commit_status` over
+    ``emit_status_transition_transactional`` and
+    :meth:`~CoordCommitRouter.commit_artifact` over ``commit_for_mission``).
+    NOT a fused ``commit()`` — the Wave-2 consumers use disjoint halves
+    (``implement.py`` = status only; ``acceptance`` = artifact only; ``move_task``
+    = both), so a single fused method is re-cut in Wave 2 (the C-006 failure).
+* **Mission-local seams** (scaffolding for ``tasks.py`` testability only; #2173
+  DROPs both, so they are explicitly NOT advertised as program-reference ports):
+  - :class:`GitOps` — porcelain reads.
+  - :class:`Render` — dual-arm output (``human`` + ``json_envelope``).
+
+Injection contract (C-005 / research D5): every orchestrator is
+``_do_<cmd>(..., *, ports: TasksPorts | None = None)``; ``None`` builds the Real
+bundle via :func:`default_ports`. The Typer ``@app.command`` never sees ``ports``
+(a Protocol-typed parameter on a decorated command collides with Typer
+introspection). This WP delivers the boundary + the FR-010 dir-equivalence proof;
+it does NOT thin the command bodies (that is WP06+).
+
+Note: ``__all__`` is intentionally omitted. This module is imported directly by
+the thin orchestrators and tests, while the former CLI-local path remains a
+compatibility shim.
+"""
+
+from __future__ import annotations
+
+import json
+import subprocess
+from collections.abc import Mapping, Sequence
+from dataclasses import dataclass
+from pathlib import Path
+from typing import Protocol, runtime_checkable
+
+from rich.console import Console
+
+from mission_runtime import MissionArtifactKind
+from specify_cli.core.commit_guard import GuardCapability
+from specify_cli.core.paths import locate_project_root
+from specify_cli.coordination.commit_router import commit_for_mission
+from specify_cli.coordination.status_transition import (
+    emit_status_transition_transactional,
+)
+from specify_cli.git.protection_policy import ProtectionPolicy
+from specify_cli.missions._read_path_resolver import (
+    _canonicalize_primary_read_handle,
+    primary_feature_dir_for_mission,
+    resolve_feature_dir_for_mission,
+    resolve_planning_read_dir,
+)
+from specify_cli.status import StatusEvent, TransitionRequest
+
+# ---------------------------------------------------------------------------
+# Mission identity value object
+# ---------------------------------------------------------------------------
+
+
+@dataclass(frozen=True)
+class MissionHandle:
+    """The two coordinates every real ``tasks.py`` seam consumes.
+
+    The canonical resolvers take ``(repo_root, mission_slug)``; this frozen pair
+    threads them through the ports as one value so the orchestrators pass a single
+    handle instead of re-plumbing both arguments at every call.
+    """
+
+    repo_root: Path
+    mission_slug: str
+
+
+# ---------------------------------------------------------------------------
+# Result value objects
+# ---------------------------------------------------------------------------
+# Renamed off ``CommitResult`` — that name is taken by ``git/commit_helpers.py``
+# (line 424); reusing it here would collide (research D10).
+
+
+@dataclass(frozen=True)
+class CommitStatusResult:
+    """Outcome of :meth:`CoordCommitRouter.commit_status`.
+
+    ``event`` is the ``StatusEvent`` the transactional emitter persisted (``None``
+    only when the transition was skipped). ``skipped`` carries the coord
+    skip-exit-0 decision so the caller reads a field instead of forking on a
+    side-effect (data-model §Ports).
+    """
+
+    event: StatusEvent | None
+    skipped: bool = False
+
+
+@dataclass(frozen=True)
+class CommitArtifactResult:
+    """Outcome of :meth:`CoordCommitRouter.commit_artifact`.
+
+    Mirrors the useful surface of ``commit_router.CommitRouterResult`` while
+    keeping the port result decoupled from that concrete type (and off the
+    ``CommitResult`` name).
+    """
+
+    status: str
+    placement_ref: str
+    commit_hash: str | None = None
+    diagnostic: str | None = None
+
+
+# ---------------------------------------------------------------------------
+# Port protocols (T008)
+# ---------------------------------------------------------------------------
+
+
+@runtime_checkable
+class FsReader(Protocol):
+    """Coordination **READ** authority (C-001, PROGRAM-REFERENCE).
+
+    Wraps the kind-aware read seam; never conflates with the write authority.
+    """
+
+    def planning_read_dir(
+        self, mission: MissionHandle, *, kind: MissionArtifactKind
+    ) -> Path:
+        """Resolve the read dir for one artifact ``kind`` (per-kind partition)."""
+        ...
+
+    def wp_tasks_dir(self, mission: MissionHandle) -> Path:
+        """Resolve the ``tasks/`` WP-prompt directory (PRIMARY partition)."""
+        ...
+
+    def primary_anchor_dir(self, mission: MissionHandle) -> Path:
+        """Resolve the topology-blind PRIMARY anchor dir.
+
+        The blind primitive ``primary_feature_dir_for_mission`` stays **inside**
+        this method, co-located with its ``_canonicalize_primary_read_handle``
+        fold (C-002).
+        """
+        ...
+
+
+@runtime_checkable
+class CoordCommitRouter(Protocol):
+    """Coordination **WRITE** authority (C-001, PROGRAM-REFERENCE).
+
+    Two capability methods over two structurally disjoint real seams
+    (research D2) — NOT a fused ``commit()``.
+    """
+
+    def feature_write_dir(self, mission: MissionHandle) -> Path:
+        """The kind-blind coord-husk write leg (``resolve_feature_dir_for_mission``)."""
+        ...
+
+    def commit_status(
+        self,
+        request: TransitionRequest,
+        *,
+        capability: GuardCapability,
+    ) -> CommitStatusResult:
+        """Commit a status transition over ``emit_status_transition_transactional``.
+
+        Keyed on :class:`GuardCapability`; self-atomic via
+        ``BookkeepingTransaction`` inside the emitter. The ``TransitionRequest``
+        is the SINGLE identity source (it already carries ``repo_root`` /
+        ``mission_slug`` / ``feature_dir`` — see the ``move_task`` call site) — so
+        this leg takes NO separate ``MissionHandle``: a second identity source
+        would be exactly the split-brain this program removes (binding refinement
+        of the indicative Phase-1 contract). The persisted ``StatusEvent`` is
+        carried back on :attr:`CommitStatusResult.event`. Used by ``implement.py``
+        and ``move_task``.
+        """
+        ...
+
+    def commit_artifact(
+        self,
+        mission: MissionHandle,
+        paths: Sequence[Path],
+        message: str,
+        *,
+        kind: MissionArtifactKind,
+        policy: ProtectionPolicy,
+    ) -> CommitArtifactResult:
+        """Commit an artifact over ``commit_for_mission`` (event-LESS).
+
+        Keyed on :class:`MissionArtifactKind` + a ``ProtectionPolicy``. Used by
+        ``acceptance`` (protected-primary routing) and ``move_task``.
+        """
+        ...
+
+
+@runtime_checkable
+class GitOps(Protocol):
+    """Porcelain reads used by ``tasks.py`` only (MISSION-LOCAL; #2173 DROP)."""
+
+    def is_dirty(self, path: Path) -> bool:
+        """True when the working tree at ``path`` has uncommitted changes."""
+        ...
+
+    def current_branch(self, path: Path) -> str:
+        """The current branch name at ``path``."""
+        ...
+
+    def is_protected(self, branch: str) -> bool:
+        """True when ``branch`` is a protected ref."""
+        ...
+
+
+@runtime_checkable
+class Render(Protocol):
+    """Output surface (MISSION-LOCAL; #2173 DROP). Dual-arm: human + JSON."""
+
+    def human(self, view: object) -> None:
+        """Render ``view`` for a human (rich tables/panels)."""
+        ...
+
+    def json_envelope(self, payload: Mapping[str, object]) -> str:
+        """Return the ``--json`` envelope string for ``payload``."""
+        ...
+
+
+# ---------------------------------------------------------------------------
+# Real adapters (T009) — wrap the canonical seams, never stubs
+# ---------------------------------------------------------------------------
+
+
+class RealFsReader:
+    """Real :class:`FsReader` over the canonical read-path resolvers."""
+
+    def planning_read_dir(
+        self, mission: MissionHandle, *, kind: MissionArtifactKind
+    ) -> Path:
+        # Annotated local: the project runs mypy with ``follow_imports = "skip"``,
+        # so the imported (typed ``-> Path``) resolver surfaces as ``Any`` here;
+        # the annotation re-pins the known concrete type without a suppression.
+        read_dir: Path = resolve_planning_read_dir(
+            mission.repo_root, mission.mission_slug, kind=kind
+        )
+        return read_dir
+
+    def wp_tasks_dir(self, mission: MissionHandle) -> Path:
+        feature_dir: Path = resolve_planning_read_dir(
+            mission.repo_root,
+            mission.mission_slug,
+            kind=MissionArtifactKind.WORK_PACKAGE_TASK,
+        )
+        return feature_dir / "tasks"
+
+    def primary_anchor_dir(self, mission: MissionHandle) -> Path:
+        # C-002: the canonicalizer fold and the blind primitive call are
+        # co-located in THIS method. The resolution-authority canonicalizer gate
+        # is intra-function (def-use); splitting the fold across the boundary
+        # turns it RED. Keep both here.
+        canonical = _canonicalize_primary_read_handle(
+            mission.repo_root, mission.mission_slug
+        )
+        anchor: Path = primary_feature_dir_for_mission(mission.repo_root, canonical)
+        return anchor
+
+
+class RealCoordCommitRouter:
+    """Real :class:`CoordCommitRouter` over the two disjoint write seams."""
+
+    def feature_write_dir(self, mission: MissionHandle) -> Path:
+        write_dir: Path = resolve_feature_dir_for_mission(
+            mission.repo_root, mission.mission_slug
+        )
+        return write_dir
+
+    def commit_status(
+        self,
+        request: TransitionRequest,
+        *,
+        capability: GuardCapability,
+    ) -> CommitStatusResult:
+        event: StatusEvent = emit_status_transition_transactional(
+            request, capability=capability
+        )
+        return CommitStatusResult(event=event, skipped=False)
+
+    def commit_artifact(
+        self,
+        mission: MissionHandle,
+        paths: Sequence[Path],
+        message: str,
+        *,
+        kind: MissionArtifactKind,
+        policy: ProtectionPolicy,
+    ) -> CommitArtifactResult:
+        result = commit_for_mission(
+            mission.repo_root,
+            mission.mission_slug,
+            tuple(paths),
+            message,
+            policy,
+            kind=kind,
+        )
+        return CommitArtifactResult(
+            status=result.status,
+            placement_ref=result.placement_ref,
+            commit_hash=result.commit_hash,
+            diagnostic=result.diagnostic,
+        )
+
+
+class RealGitOps:
+    """Real :class:`GitOps` over ``git`` porcelain + the resolved protection policy."""
+
+    def __init__(self, repo_root: Path | None = None) -> None:
+        self._repo_root = repo_root or locate_project_root() or Path.cwd()
+
+    def is_dirty(self, path: Path) -> bool:
+        completed = subprocess.run(
+            ["git", "-C", str(path), "status", "--porcelain"],
+            capture_output=True,
+            text=True,
+            check=False,
+        )
+        return bool(completed.stdout.strip())
+
+    def current_branch(self, path: Path) -> str:
+        completed = subprocess.run(
+            ["git", "-C", str(path), "rev-parse", "--abbrev-ref", "HEAD"],
+            capture_output=True,
+            text=True,
+            check=False,
+        )
+        return completed.stdout.strip()
+
+    def is_protected(self, branch: str) -> bool:
+        protected: bool = ProtectionPolicy.resolve(self._repo_root).is_protected(branch)
+        return protected
+
+
+class RealRender:
+    """Real :class:`Render` over rich + ``json.dumps`` (default separators)."""
+
+    def __init__(self, console: Console | None = None) -> None:
+        self._console = console or Console()
+
+    def human(self, view: object) -> None:
+        self._console.print(view)
+
+    def json_envelope(self, payload: Mapping[str, object]) -> str:
+        # Default separators — byte-identical to the inline ``json.dumps(data)``
+        # sites in tasks.py (research D10; parity obligation for Render).
+        return json.dumps(payload)
+
+
+# ---------------------------------------------------------------------------
+# Bundle + Real builder (T011)
+# ---------------------------------------------------------------------------
+
+
+@dataclass(frozen=True)
+class TasksPorts:
+    """The injected capability bundle (exactly four ports)."""
+
+    fs: FsReader
+    coord: CoordCommitRouter
+    git: GitOps
+    render: Render
+
+
+def default_ports() -> TasksPorts:
+    """Build the production (Real-adapter) bundle.
+
+    This is what an orchestrator constructs when ``ports is None`` at the
+    ``_do_<cmd>(*, ports=None)`` injection seam (C-005).
+    """
+    return TasksPorts(
+        fs=RealFsReader(),
+        coord=RealCoordCommitRouter(),
+        git=RealGitOps(),
+        render=RealRender(),
+    )
