+++ tests/architectural/resolution_gate_allowlist.yaml
@@ -115,11 +115,11 @@ coord_authority:
   # removed — WP08 (2dda443d5) folded its probe to primary; merged list_dependents
   # has zero resolve calls.
   - qualname: list_tasks
-    line: 2757
-    rationale: "pre-sweep baseline write (src/specify_cli/cli/commands/agent/tasks.py) - STATUS-partition coord-aware event-log read (_lt_feature_dir); tasks-py-degod WP09 re-pin 2198->2746, then +11 to 2757 (move_task #1386 protected-branch guard-order early-exit added lines above; same site, not drained)"
+    line: 2768
+    rationale: "pre-sweep baseline write (src/specify_cli/cli/commands/agent/tasks.py) - STATUS-partition coord-aware event-log read (_lt_feature_dir); tasks-py-degod WP09 re-pin 2198->2746, then +11 to 2757 (move_task #1386 protected-branch guard-order early-exit), then +11 to 2768 (self-review precedence guard; same site, not drained)"
   - qualname: validate_workflow
-    line: 3825
-    rationale: "pre-sweep baseline write (src/specify_cli/cli/commands/agent/tasks.py) - STATUS-partition coord-aware event-log read (_vw_feature_dir); tasks-py-degod WP09 re-pin 2995->3814, then +11 to 3825 (move_task #1386 protected-branch guard-order early-exit added lines above; same site, not drained)"
+    line: 3836
+    rationale: "pre-sweep baseline write (src/specify_cli/cli/commands/agent/tasks.py) - STATUS-partition coord-aware event-log read (_vw_feature_dir); tasks-py-degod WP09 re-pin 2995->3814, then +11 to 3825 (move_task #1386 protected-branch guard-order early-exit), then +11 to 3836 (self-review precedence guard; same site, not drained)"
   - qualname: implement
     line: 1403
     rationale: "pre-sweep baseline write (src/specify_cli/cli/commands/agent/workflow.py) - coord-read-residuals WP01 line-drift re-pin 1395->1403 (identity routing added lines above)"
